Refactored folder structure of card components
Extracted styles to associated components.
